# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: CUENTY API - Enterprise
  description: |
    API REST completa para el sistema CUENTY de gestión de suscripciones para el mercado mexicano.
    
    ## Características
    - Autenticación JWT + API Keys
    - Rate limiting: 100 requests/minuto
    - Webhooks para integraciones
    - SDKs disponibles: Node.js, Python, PHP
    
    ## Ambiente
    - Base URL Producción: https://api.cuenty.com/v1
    - Base URL Sandbox: https://sandbox.api.cuenty.com/v1
    
  version: 3.0.0
  contact:
    name: CUENTY API Support
    email: api@cuenty.com
    url: https://cuenty.com/docs
  license:
    name: Propietario
    url: https://cuenty.com/legal/api-terms
servers:
  - url: https://api.cuenty.com/v1
    description: Servidor de Producción
  - url: https://sandbox.api.cuenty.com/v1
    description: Servidor de Sandbox
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para desarrolladores externos
  
  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
...
paths:
...
  /payments/spei/generate:
    post:
      tags:
        - Payments
      summary: Generar referencia SPEI
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orderId
                - amount
              properties:
                orderId:
                  type: integer
                amount:
                  type: number
                  format: decimal
      responses:
        '200':
          description: Referencia SPEI generada
          content:
            application/json:
              schema:
                type: object
                properties:
                  reference:
                    type: string
                  clabe:
                    type: string
                  amount:
                    type: number
                  expiresAt:
                    type: string
                    format: date-time
  /payments/codi/generate:
    post:
      tags:
        - Payments
      summary: Generar QR CoDi
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orderId
                - amount
              properties:
                orderId:
                  type: integer
                amount:
                  type: number
      responses:
        '200':
          description: QR CoDi generado
          content:
            application/json:
              schema:
                type: object
                properties:
                  qrCode:
                    type: string
                    description: Base64 encoded QR image
                  codiReference:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
  
  # ==========================================================================
  # ANALYTICS
  # ==========================================================================
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []